refactor: component 패키지 삭제
