+++ tomcat/src/main/java/org/apache/coyote/http11/request/RequestUri.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.request.component;
+package org.apache.coyote.http11.request;
 
 import java.util.Arrays;
 import java.util.List;
